共通レイアウトを追加
