+++ src/App.tsx
@@ -1,17 +1,22 @@
 import React from 'react'
 import { Route, Switch } from 'react-router-dom'
 import styled from 'styled-components'
+import Layout from './components/Layout/Layout'
 import Login from './pages/Login/Login'
 import Search from './pages/Search/Search'
 
 function App(): JSX.Element {
   return (
     <RootDiv>
-      <header className="App-header"></header>
-      <Switch>
-        <Route exact path="/" component={Search} />
-        <Route exact path="/login" component={Login} />
-      </Switch>
+      <Layout>
+        <>
+          <header className="App-header"></header>
+          <Switch>
+            <Route exact path="/" component={Search} />
+            <Route exact path="/login" component={Login} />
+          </Switch>
+        </>
+      </Layout>
     </RootDiv>
   )
 }
